Amélioration design + tranfert parties enregistrées par bluetooth

+++ app/src/main/java/com/tambyy/fanoronaakalana/graphics/drawable/item/Akalana.java
@@ -25,7 +25,7 @@ public class Akalana extends Drawable {
         this.paint.setColor(color);
     }
 
-    public void setLineWidth(int lineWidth) {
+    public void setLineWidth(float lineWidth) {
         this.paint.setStrokeWidth(lineWidth);
     }
